pyproject.toml
[project]
name = "autobump"
version = "0.1.0"
authors = [{ name = "Liam", email = "info@liampower.dev" }]
description = "Automatically bump version numbers for your project"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "gitpython>=3.1.44",
    "packaging>=25.0",
    "semver>=3.0.4",
    "tomlkit>=0.13.2",
    "typer>=0.15.3",
]

[project.scripts]
autobump = "autobump.cli:app"
